requirements.txt
joblib==0.15.1
numpy==1.19.0
opencv-python==4.2.0.34
Pillow==7.1.2
scikit-learn==0.23.1
scipy==1.5.0
sklearn==0.0
threadpoolctl==2.1.0
